# libfatx: Use writeback cache for fat entries