Implements a writeback cache for fat entries. This has a significant performance impact across all filesystem operations.

Below are the results of a simple RW benchmark. The drive image used is a clean copy of the extracted `./tests/xbox_hdd.img` image. No arguments were passed to fatxfs besides the image and mount point.

`master` without caching:
```
> dd if=/dev/zero of=test bs=64k count=4k conv=fdatasync
4096+0 records in
4096+0 records out
268435456 bytes (268 MB, 256 MiB) copied, 103.525 s, 2.6 MB/s
> dd if=test of=/dev/null bs=64k count=4k               
4096+0 records in
4096+0 records out
268435456 bytes (268 MB, 256 MiB) copied, 2.63456 s, 102 MB/s
```

With caching:
```
> dd if=/dev/zero of=test bs=64k count=4k conv=fdatasync
4096+0 records in
4096+0 records out
268435456 bytes (268 MB, 256 MiB) copied, 13.7842 s, 19.5 MB/s
> dd if=test of=/dev/null bs=64k count=4k               
4096+0 records in
4096+0 records out
268435456 bytes (268 MB, 256 MiB) copied, 0.386977 s, 694 MB/s
```

`master` with `big_writes` FUSE option:
```
> dd if=/dev/zero of=test bs=64k count=4k conv=fdatasync
4096+0 records in
4096+0 records out
268435456 bytes (268 MB, 256 MiB) copied, 25.6293 s, 10.5 MB/s
> dd if=test of=/dev/null bs=64k count=4k               
4096+0 records in
4096+0 records out
268435456 bytes (268 MB, 256 MiB) copied, 2.66338 s, 101 MB/s
```

Caching with `big_writes` FUSE option:
```
> dd if=/dev/zero of=test bs=64k count=4k conv=fdatasync
4096+0 records in
4096+0 records out
268435456 bytes (268 MB, 256 MiB) copied, 3.08812 s, 86.9 MB/s
> dd if=test of=/dev/null bs=64k count=4k               
4096+0 records in
4096+0 records out
268435456 bytes (268 MB, 256 MiB) copied, 0.389961 s, 688 MB/s
```